# fix compare script - output formatting - correctly align numbers >9999