Fix output formatting of compare_bench.py , correctly align numbers greater than 9999.

Before (numbers >9999 in Old and New columns are incorrectly aligned):
```
Benchmark                                                                 Time           CPU           Old           New
------------------------------------------------------------------------------------------------------------------------
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx1/threads:1                             -0.92         -0.92         1234          103
BM_xxxxxxxxxxxxxxxxxxxxxxxxxx2/threads:1                                 -0.11         -0.10         29096         26108
BM_xxxxxxxxxxxxxxxxxxxx3/threads:1                                       -0.20         -0.20         21060         16759
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx4/threads:1                        +6.77         +6.79          415         3234
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx5/threads:1                            -0.19         -0.19         54845         44358
BM_xxxxxxxxxxxxxxxxxxxxxxxxx6/threads:1                                  -0.22         -0.22         131896         103098
```

After:
```
Benchmark                                                                 Time           CPU           Old           New
------------------------------------------------------------------------------------------------------------------------
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx1/threads:1                             -0.92         -0.92          1234           103
BM_xxxxxxxxxxxxxxxxxxxxxxxxxx2/threads:1                                 -0.11         -0.10         29096         26108
BM_xxxxxxxxxxxxxxxxxxxx3/threads:1                                       -0.20         -0.20         21060         16759
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx4/threads:1                        +6.77         +6.79           415          3234
BM_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx5/threads:1                            -0.19         -0.19         54845         44358
BM_xxxxxxxxxxxxxxxxxxxxxxxxx6/threads:1                                  -0.22         -0.22        131896        103098
```