@EricWF 2016-12-09
Please run the tests using `python report.py`. They are currently failing for me and are outputting very poorly aligned output.

I suspect that fixing the alignment is going to take a little more work. I really phoned in the original code.