@pcampr 2016-12-09
I run the tests using `python report.py`, which failed even in the master version. I propose a fix for the test (two new testing benchmarks with larger numbers added), and updated formatting string.

Before (master version with updated test, but not changed formatting string):
```
$ python report.py
Benchmark                   Time           CPU           Old           New
--------------------------------------------------------------------------
BM_SameTimes               +0.00         +0.00           10           10
BM_2xFaster                -0.50         -0.50           50           25
BM_2xSlower                +1.00         +1.00           50          100
BM_10PercentFaster         -0.10         -0.10          100           90
BM_10PercentSlower         +0.10         +0.10          100          110
BM_100xSlower              +99.00         +99.00          100         10000
BM_100xFaster              -0.99         -0.99         10000          100
.
----------------------------------------------------------------------
Ran 1 test in 0.002s

OK
```

After (this PR version):
```
$ python report.py
Benchmark                   Time           CPU           Old           New
--------------------------------------------------------------------------
BM_SameTimes               +0.00         +0.00            10            10
BM_2xFaster                -0.50         -0.50            50            25
BM_2xSlower                +1.00         +1.00            50           100
BM_10PercentFaster         -0.10         -0.10           100            90
BM_10PercentSlower         +0.10         +0.10           100           110
BM_100xSlower             +99.00        +99.00           100         10000
BM_100xFaster              -0.99         -0.99         10000           100
.
----------------------------------------------------------------------
Ran 1 test in 0.002s

OK
```